+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Shape::rename()` to rename a key in the output array
 - `Shape::default()` to specify a default value when an optional key is not set
+- `Is::just()`
 
 ## 1.5.0 - 2024-11-10