+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.6.0 - 2024-11-11
+
+### Added
+
+- `Shape::rename()` to rename a key in the output array
+- `Shape::default()` to specify a default value when an optional key is not set
+- `Is::just()`
+
 ## 1.5.0 - 2024-11-10
 
 ### Added